starting to work on about page
