+++ www/src/components/layouts/common/Header/items.ts
@@ -1,0 +1,22 @@
+// TODO: replace soon hrefs as features get built out
+// TODO: replace this with a call to gatsby to get config based values
+const MENU_ITEMS = [
+  {
+    title: `About`,
+    href: `/about`,
+  },
+  {
+    title: `Writing`,
+    href: `/404`,
+  },
+  {
+    title: `Portfolio`,
+    href: `/404`,
+  },
+  {
+    title: `Contact`,
+    href: `/404`,
+  },
+];
+
+export default MENU_ITEMS;
